[SKIP]: Explain the repository in the README.md
Explain that the repository is a collection of demonstrations of how to
implement features by sending [HTML over the
wire](https://hotwired.dev).

How to read this repository
---

Through the power of incremental Git diffs, each of this repository's
[branches][] provides a step-by-step demonstration of how to implement a
feature or behavior.

This repository's [main][] branch serves as the root all of the other
branches, and consists of a handful of commits generated by the Rails
command line interface.

When reading a branch's source code, read the changes commit-by-commit
either on the branch comparison page (for example,
[main...hotwire-example-live-preview][]) or the branch's commits page
(for example, [hotwire-example-live-preview][]).

[branches]: https://github.com/thoughtbot/hotwire-example-template/branches
[main]: https://github.com/thoughtbot/hotwire-example-template/tree/main
[main...hotwire-example-live-preview]: https://github.com/thoughtbot/hotwire-example-template/compare/hotwire-example-live-preview
[hotwire-example-live-preview]: https://github.com/thoughtbot/hotwire-example-template/commits/hotwire-example-live-preview